'time' notion based on millis() and watchdog (during low-power sleep) --> too complex --> need to refactor

+++ io_util.h
@@ -9,7 +9,7 @@
   typedef uint32_t time32_ms_t;
   typedef uint16_t time16_s_t;
 
-  typedef int32_t duration16_ms_t;
+  typedef int16_t duration16_ms_t;
   typedef int32_t duration32_ms_t;
   
   typedef uint16_t millivolt_t;
